+++ moor/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## unreleased
 
 - Fix crash when `customStatement` is the first operation used on a database ([#199](https://github.com/simolus3/moor/issues/199))
+- Allow transactions inside a `beforeOpen` callback
 
 ## 2.0.1